Refactor sol-cov to de-duplicate code for coverage and profiling
